Address review feedback: trim comment/NEWS, uint8_t running, add func-path reentrancy test
Per brijkapadia's review on GH-154571:
- Remove the explanatory comment above test_accumulate_reenter.
- Change accumulateobject.running from int to uint8_t (no call sites
  needed changes; assigning 0/1 to a uint8_t requires no source edit).
- Shorten the NEWS entry, keeping one clause on what was actually
  broken (silent corruption) rather than only the new behavior.

Also add test_accumulate_reenter_func, covering reentrancy triggered
via the func callable rather than the source iterable -- the guard
protects both paths but only the iterable path had a test.

teedataobject.running (same file, same purpose, for itertools.tee) is
deliberately left as int rather than also narrowed, to avoid scope
creep beyond what was requested.

+++ Modules/itertoolsmodule.c
@@ -3079,7 +3079,7 @@ typedef struct {
     PyObject *binop;
     PyObject *initial;
     itertools_state *state;
-    int running;
+    uint8_t running;
 } accumulateobject;
 
 #define accumulateobject_CAST(op)   ((accumulateobject *)(op))
